Refactor `RootPath` into dedicated module
