+++ src/storage/mod.rs
@@ -4,8 +4,10 @@ mod group;
 mod logging;
 mod persistent;
 mod directory;
+mod root;
 
 pub use group::Group;
 pub use logging::*;
 pub use persistent::{Persistent, FILETYPE};
 pub use directory::*;
+pub use root::*;
